requirements.txt
certifi==2019.11.28
chardet==3.0.4
Deprecated==1.2.7
idna==2.8
invoke==1.4.1
PyGithub==1.45
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
six==1.14.0
tqdm==4.42.1
urllib3==1.25.8
wrapt==1.11.2
